refactor: extract redirect handling
